Always kill dangling process
It is possible that subprocess.wait returns without actually killing the
process, resulting in dangling processes.

For example it is possible that `adb wait-for-device` does not properly
get killed in mac os. This in turns can result in dangling threads that
are logging the output of the process.

We now always terminate a process if we think it is still alive.

Change-Id: Icd6036e6c51a7e3e141fa549629c72093c0deed6