Add e2e test validating input events
This adds an e2e test that verifies that the animation app is receiving
all the expected events we send via gRPC.

Note this approach is a bit limited:

- We can only detect a combination of up/down events
- We have to do some rate limiting

Since we have -debug-events enabled again we re-enable the tests
based on validating the arrival of events.

The animation app will now mark itself as started once it has rendered
a series of frames.

Note: This upgrades the apk to be signed, as we can no longer check in
debug apks.

Bug: 245964307
Change-Id: I53a7d895cb637127dd922742d3640785b614ceee