Update boot tests with notification
This adds an integration tests that validates we now deliver boot
notifications over gRPC

It also fixes as subtle bug in the wait_until function where we secretly
assumed that if a predicate returns true, it will always return
true.

This was due to the double invocation during exit, which could make it
possible for this to return False, even though we didn't timeout yet.

Change-Id: I4bcd927bc86430077d94f888bc8fbba734b3a49b