Enable support for Wear images
This fixes a small issue where we did not recognize wear based system
images. Wear images are only available in x86 abi, so we have to
explicitly define which abi we want to use for the image.

Note: You probably do not want to run these on arm based macs.

Includes a sample test suite for wear images that can be used to develop
a set of unit tests.

Test:
  ./run_tests.sh --test_config cfg/wear_x86_tests.json -e \
  ~/src/emu-master-dev/external/qemu/objs/emulator
Change-Id: I7640192ddd1073eacb46a3344dc4972c597320be