Add support for using mobly within pytest
This adds support and two samples of how to use mobly to interact with
a running emulator.

Mobly will allow you to write tests using snippets that run on android
and are controlled from pytest.

See the documentation for more details on how to write mobly tests.

This updates the AnimateBox app to include a mobly snippet and migrates
the build scripts to the proper gradle version. This will likely require
a java jdk upgrade.

Bug: 302605853
Change-Id: Ic04ef603f18c050e9a15b7ce6092788f5936193e