Weaken launch assertions
The way we check for the launch of an activity does not always work
properly. Instead of checking for the window in focus we now check if
the activity is in the set of running activities

We also include the logging of the individual stages of the fixtures to
make it easier to figure out which fixture is doing what during setup
and teardown.

Bug: 303058205
Change-Id: I6dd66e34d0732e11a4b4f5e1f57c32cf98300ced