Reimplement E2E tests using asyncio
We are seeing a series of locks in our tests that are caused by us
managing threads ourselves. Instead of managing threads ourselves we now
switch to using pythons asyncio framework

This allows us to write the tests with the test framework explicitly
being aware of the concurrency. It also enables us to offload all the
concurrency to python directly.

- No more process and I/O management for logs
- We can now use `asyncio.wait_for` to properly create timeouts
- We no longer have to manager our own locks
- No more "hangs" when the emulator crashes (gRPC connection failures
  are now immediately detected)

Hopefully this will reduce the likelyhood of hanging threads due to
emulator/test issues

The README.MD contains an updated section on what asyncio is, why we
adopted it and how you can use it.

Bug: 308938749
Change-Id: I78a3aecc5d36e946665100c4bd580e02f47fe103