Use keyboard input to advance testGoogleLoginSettings.
Use that keyboard to advance the test, instead of the Next button as the
keyboard covers the Next button on smaller screens. The test will click
the Enter key on the keyboard twice, once after the email input, and
once after the password.  The Next button will not be used.

BUGS: b/320526624
Change-Id: I803b3919c785d2757d6287c6d3e39f20734dec2d