Reduce logging spam and add missing await
We use `logcat -d` to make sure we have newer loglines after clearing
the log. This can result in excessive logging if the available number of
logs are large.

We now take the tail, so we do not introduce log spam and include the
await statement so we are properly waiting for the logcat to be cleared.

Change-Id: I8e3af2600082f1fb9a389464476c2c71effe7aee