Flush buffers on write
It appears that some of the console tests do not always succeed. One
possibility is that we do not complete write all the bytes:

See this log:

14:13:04.487 34_google_apis_arm64-v8a_Pixel2-con      INFO    --> rotate
14:13:04.487 34_google_apis_arm64-v8a_Pixel2          INFO    in control_client read: received rotate
14:13:04.487 34_google_apis_arm64-v8a_Pixel2          INFO
14:13:04.487 34_google_apis_arm64-v8a_Pixel2          INFO    in control_client read: size < 0, exiting with 35: Resource temporarily unavailable

It could be that we did not write the '\n' character, causing us to
timeout.

We now forcefully drain all the write queues. This makes sure we do not
have any bytes lingering in sockets.

Change-Id: I42da56507f4bf6ca2c97b0a422f79f573720aac1