Insert waitUntilGone for email and password links in loginGoogleApp.
The test was looking for the Password link while still on the Email
login page. Add a couple of waitUntilGone() methods to
GoogleAppUtil.loginGoogleApp(). This change will allow the test will wait
for the email settings page to be gone before looking for password info,
and wait for the password page to be gone before continuing.

BUGS: b/322202670
Change-Id: I27ca46ec63bb5a32be366aee6e417dbde8150a23