Create launchPath_v2 as API 31 fallback for failed scrolling.
There is a special case logic in the `appLaunch` file exclusive to API
31 that improves overall UI scrolling results when added. The scrolling
issue in API 31 is not consistent and may be due to the inconsistent
underlying XML.

This change creates a new method, which is for special cases in API 31
dealing with the scrolling logic. It's not called everytime, instead it
only severes as a fallback case if the original appLaunch fails in API
31. The two methods work togeter to try both application scrolling
options.

Make `appLaunch` more readable by stripping out API 31 specific code and
allow API 31 tests to try scrolling either way before the test fails due
to flaky UI scrolling logic.

BUG: b/328433515
Change-Id: I64388679126c1d4213e70fa0dc3c0e43c3ba9308