Fix missing gradle XML reports for Sys Img UI tests.
In current gradle version, the path to the instrumented tests' XML reports has an extra 'debug' folder appended, which cause the reports to not be found and therefore not copied to the final test logs. This fix changes the procedure of finding the xml files by allowing the existence of a subdir in the gradle (test) output folder.

 - Fix os name syntax (==) warning.

 - Fix gradle XML report path to include the full path of the gradle xml output report
 when copying to the testlogs.

BUG: b/327455714
Change-Id: I26b13b98170889155bf92e912645a311a3ae8cae