Restart crashed emulator
A crash in the middle of a module can cause multiple tests in the same
module to fail. To prevent this we now rewrite the avd fixture as
a function level fixture that will validate that the emulator is
running, and restart it if that is not the case.

This will make sure a crash will only result in a single test failure
versus failing all the tests in the module.

Change-Id: I1a69916eeacc1d6dc4570a6ba0c4833803c8acd4