Support using fetcher to get system images.
The fetcher tool is used by the worker to fetch the emulator and
symbols files before running pytest. It has built in support for
caching and removing old artifacts.

Use of the tool would allow better control over disk space usage
for workers with smaller SSDs, as well as opening the door for
pytests to use TOT builds. It is also a step towards removing
the need for the full emulator source tree to be checked out to
run the tests.

Bug: 356846072
Change-Id: I2800d0eba2d8f4c431291664f02ee6941a4c67b5