Add support for presubmit
This enables the running of presubmit checks when you run runtest.sh
with the --presubmit flag.

The presubmit flag takes either a commit hash, or a filename, and will
restrict the test suite to only run the tests that are defined in the
test file.

For example:

./run_tests.sh -e ~/src/emu-master-dev/external/qemu/objs/emulator
--presubmit HEAD HEAD~1

will run all the tests that are modified in the commit at HEAD and
HEAD~1

./run_tests.sh -e ~/src/emu-master-dev/external/qemu/objs/emulator
--presubmit pytest/test_embedded/tests/sms/test_sms.py tests/test_boot.py

Will restrict the test suite to only run tests in test_sms.py and test_boot.py

This can be used:

- By developers to make sure they run all the test before they submit
  (--presubmit HEAD)

- Build bots to run the tests on changed files

Bug: 371344784
Change-Id: I021a7352706d38767ed575b746e3d663a58928ec