Add screen_recorder fixture
This adds a screen recorder fixture that will record the screen. This
can be useful for debugging UI tests. The addition of the fixture to
your test will record an mp4 under:

screenrecording/name_of_your_test.mp4

The screen recorder fixture makes use of the mss library to do fast
screen capturing and the opencv library to do the video encoding.

This includes the dependencies needed to make use of screenrecording.

Bug: 365169284
Test: A recording of test_close_emulator was created.
Change-Id: Ifc5850b6696ba475c3659d26f522fcb0be727319