Account for time drift in screen recorder
The simple screeen recorder could fall behind due to:

- Time taken to take a screenshot.
- Time drift in asyncio.sleep

Which resulted in videos that would go really fast! We now adjust for
this by calculating time drift and writing out additional duplicate frames
to make sure we keep up with the framerate.

We also enable the recording for any type of streaming, reducing the
number of image we are saving everywhere.

Change-Id: I51ab6a976b8f4061e009d89a5748c4385f6d7593