Run Black in a temporary virtual environment
This commit introduces the ability to run the Black code formatter in
a temporary virtual environment. This ensures that Black is always
available, even if it's not installed globally on the system.

This will make sure the repo commit hooks can successfully run, even
when black is not pre-installed on the system.

Change-Id: I6ee3a1388221aac664d47d7709b516b9b385bbde