Fix the path to the QR code image for qrcode_png and qrcode_mp4.
These fixtures used to use the conftest module path to locate the QR
code PNG image under the cfg folder. The fixtures were then moved to a
new folder. This change corrects the path where the PNG image is
expected to be. Add an assertion test to detect similar path errors.

BUG: b/373477391
Change-Id: I0039a8ba757807baf26a97ced8c91c8f71f15d41