Reduce default logging levels
This change reduces the default logging levels for several
commands and modules to reduce noise in the logs.

The following commands are affected:

- create_zip.py
- kill_emulator.py
- run_tests.py
- test_runner.py

The following modules are affected:

- emu.logging.log_handler
- emu.process.kill_emulator

The `--log-level` option can be used to override the default
logging level. The `-v` and `--verbose` options have been
removed from the help text, but they can still be used to
enable verbose logging.

Note: This only impacts the logging of the output of the test runner,
the individual tests will still log everything to the log files.

Change-Id: I61409e843603ba3743d6d32a1d4b6897ec66d224