Write adb logcat to file only
Prevent logcat from writing to stdout, instead redirecting it to a file
on disk. This prevents the build.log from being polluted with
excessive logcat output (over 400mb).

The location of the logcat file will be logged to the console. It can
typically be found in the testlogs/ or logs/ directory on build bots
and will be included in the pytest test result tarball.

Bug: 382107161
Change-Id: I62e934e197dc6d45731c2b7d2dc0f5ec01e36f97