fix: Update wait_until to timeout immediately
The wait_until function has been updated to timeout immediately and only
accept async functions. This change is necessary because we have moved
to async functions a while back and no longer need to support synchronous
routines.

Additionally, typing information has been added to the timing functions
to detect incorrect usage and assist with VS Code's IntelliSense.

Note: We cannot guarantee this timeout when using synchronous functions,
in that case a warning will be issued.

Bug: 385348775
Change-Id: I63ab17ea2367462503ce5caedd2f707801fbfb82