test: Refactor test_emulator_controls_keys for embedded
This commit refactors the test_emulator_controls_keys test for embedded emulators.

The test was previously a single monolithic test consisting of 7 subtests.
This commit separates the test into subtests in their own file and uses fixtures
to ensure preconditions are met for every individual test. It also
removes the back button dependency on the message app, which behaves
differently on various API levels. Instead, the animation app is now
used to check the back stack of Android.

Bug: 385389033
Change-Id: I8377d99717ea92d749197da02dcb4a20fb57b35e