test: Compare all volume levels in audio test
The audio volume test used to check only the music stream.
However, it is possible that the volume button would change other streams.

Instead of only looking at the audio stream we look at the vector of all
audio levels and we use mobly to reset the volumes to a known state.

Bug: 385498968
Change-Id: Id2701868098ffc5c219aba2b5d5ed49bb364a4ae