test: Disable PIL and PyScreeze screen grabbers
Both PIL and PyScreeze are extremely slow when it comes to capturing
frames, negatively impacting the tests we run.

We now will only use the mss strategy and fall back to none if we are
unable to obtain it.

It can sometimes around 1s to grab a single frame with PIL and PyScreeze,
increasing flakiness.

Change-Id: I43136b7be6299c8c409799de643d7024f970ab60