test: Do not reset emulator state for dependent tests
This change allows defining test dependencies that leave the emulator
state as is. This ensures that one test can create a state consumed by
a dependent test.

For example:

test 1: Can bring up the power-menu
test 2: Clicks a button in the power menu

Includes a set of tests to validate the behavior and updates the
documentation to explain the new behavior.

Change-Id: I914dbecbc921edcc8cc998c67fa66fb2a759a877