refactor: Remove timeouts from snapshot download tests
This refactors the snapshot download tests to remove all direct
dependencies on timeouts. It uses `requests` for downloads with
error handling and `eventually` for checking emulator status. This
makes the tests more robust and less prone to timing issues.

Note: The test will consistently fail due snapshots not being
compatible.

Bug: 375280350
Test: pytest -vvv \
  --emulator=$HOME/src/emu-master-dev/external/qemu/objs/distribution/emulator/emulator
  -k test_can_load_oldsnapshot
Change-Id: Ie72f1896ca72adb607b20e906efc400fcb000d11