Optimize camera handling for test_ar_sanity.
Split the 'camera_activity' fixture into two. 'camera_app' is a new
fixture that provides an instance of the 'CameraApplication' subclass.
'camera_ready' is a fixture that dismisses unexpected dialogs and
ensures the back camera is ready for use.

Add the 'camera_is_active' function to ensure the back camera is
activated when unexpected dialogs appear.

Add a watcher to trigger an emulator exception if the "System UI isn't
responding" dialog is detected.

BUG: b/396157697
Change-Id: Ic2a7e323bc16ac5cbb8c22eb8a73089289fbb56e