Update to allow pre-extraction of archives.
Now that the flow is established, the code is updated to allow for
using pre-extracted archives. This moves the extraction to the host
with the important benefit that the time is not part of the test
timeout.

This also switches to gcs_archives like the cts tests, but uses a
generic BUILD file contents for simplicity. It also uses the location
of said BUILD file so that all paths are to the parent directory
instead of being at different levels.

Bug: 454698684
Change-Id: Ib4063f421ecfeda6416d9541144b8b76ff9cfb7e