Update ets to no longer use a module extension.
This was originally done as the goldfish code was not yet modularized
and an extension was the only approach. This should simplify the code
and eliminate the need for the extra .bzl logic.

Since the build.py file needed to be updated anyways, the ets
directory is moved up a level as the tests parent directory was making
paths longer without adding value.

Bug: 454698684
Change-Id: Iad0ff91ea18d59f1105291dbf100be70a288c6f5