Rework ets config to be more programmatic.
The text templates were always intended as a stop-gap to get the full
process running until a new re-usable config was in place. This CL
prepares for using python based protos as the config by inverting the
logic so the config is the main file, calling the runner as a library.

Another major change was dropping the handy bzl manipulation of
data dependencies and $(location) maps. While this made the rules
easier to write, select() statements could not be used so that approach
would never work across multiple OSes.

Lastly a _print test was added which will simply print the config so
.txtpb can be manually inspected.

Bug: 454698684
Change-Id: I9219921694492338f1c58e1a649c2ac6bafc6c5e